refactor: extract shared test utility in brainstorm-server tests (#56)
Consolidate duplicated `test` function logic across `helpers.test.js`,
`server.test.js`, and `ws-protocol.test.js` into a new shared helper
`test-utils.js`.

The new `createSuite` utility supports both synchronous and asynchronous
(Promise-returning) tests, which were used across the different files.
It also provides a standardized `summary` reporting mechanism.
